refactor(cli): flatten whoami command module
